requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
eval_type_backport==0.3.1
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
invoke==2.2.1
jiter==0.13.0
mistralai==1.12.0
mkl_fft==2.1.1
mkl_random==1.3.0
mkl-service==2.5.2
numpy==2.4.2
openai==1.59.7
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pandas==3.0.0
pip==25.3
protobuf==6.33.5
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.3
setuptools==80.10.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
tqdm==4.67.3
tree-sitter==0.21.2
tree-sitter-cpp==0.21.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
wheel==0.46.3
zipp==3.23.0
